refactor(matcha-window): split `runtime` into native/web modules
Moves the previously inline `mod native { ... }` / `mod wasm { ... }`
bodies into dedicated `runtime/native.rs` and `runtime/web.rs`,
mounted via `#[path]`. Each backend file is cfg-free.
